+++ CMakeVehicles/AirVehicles/Eagle.cpp
@@ -1,7 +1,7 @@
 #include "Eagle.h"
 
 namespace DynamicVehicles {
-	Eagle::Eagle(int distance_) {
+	Eagle::Eagle(float distance_) {
 		nameVehicle = "Îð¸ë";
 		velocity = 8;
 		distanceReductionFactor = 0.94;
