+++ CMakeVehicles/GraundVehicles/Camel.h
@@ -4,8 +4,8 @@
 namespace DynamicVehicles {
 	class Camel : public DynamicVehicles::GraundVehicles {
 	public:
-		Camel(int distance_);
-		VEHICLESLIBRARY_API int calculateTime() override;
+		Camel(float distance_);
+		VEHICLESLIBRARY_API float calculateTime() override;
 	};
